requirements.txt
django-phone-field==1.8.1
Django==3.0.8
freezegun==0.3.15
python-dateutil~=2.8.1
selenium~=3.141.0